Add wp_auth0_get_option function to get option values
